AkonadiDataSourceQueries: Use a single query for the sources tree.
The recursive datasource queries are mapped to a single tree query that
is build on top of a cache.

AkonadiCollectionTreeSource now directly models this tree query, that
could ideally be replaced directly with a query provided by the akonadi
framework.

On top of that sits a query multiplexer that multiplexes the updates to
the various sub-queries (that's kind of the workaround until we have a
proper tree query).